requirements.txt
requests~=2.31.0
python-dotenv~=1.0.0
google-api-python-client==2.113.0
pika==1.3.2
